add stats module

+++ src/modules/label/label.service.ts
@@ -1,9 +1,9 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
-import { CreateLabelDto } from './dto/create-label.dto';
-import { UpdateLabelDto } from './dto/update-label.dto';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
-import { Label } from './entities/label.entity';
+import { Injectable, NotFoundException } from '@nestjs/common'
+import { CreateLabelDto } from './dto/create-label.dto'
+import { UpdateLabelDto } from './dto/update-label.dto'
+import { InjectRepository } from '@nestjs/typeorm'
+import { Repository } from 'typeorm'
+import { Label } from './entities/label.entity'
 
 @Injectable()
 export class LabelService {
@@ -13,30 +13,50 @@ export class LabelService {
   ) {}
 
   async create(createLabelDto: CreateLabelDto) {
-    const label = this.labelRepository.create(createLabelDto);
-    return this.labelRepository.save(label);
+    const label = this.labelRepository.create(createLabelDto)
+    return this.labelRepository.save(label)
   }
 
-  findAll() {
-    return this.labelRepository.find();
+  async findAll() {
+    const labels = await this.labelRepository.find()
+    return Promise.all(
+      labels.map(async (label) => {
+        const result = await this.labelRepository
+          .createQueryBuilder('label')
+          .leftJoinAndSelect('label.scenes', 'scene')
+          .where('label.id = :id', { id: label.id })
+          .getOne()
+        return { ...label, sceneCount: result?.scenes?.length || 0 }
+      }),
+    )
   }
 
   async findOne(id: string) {
-    const label = await this.labelRepository.findOne({ where: { id } });
+    const label = await this.labelRepository.findOne({ 
+      where: { id },
+      relations: ['scenes']
+    })
     if (!label) {
-      throw new NotFoundException(`Label with ID ${id} not found`);
+      throw new NotFoundException(`Label with ID ${id} not found`)
     }
-    return label;
+
+    return { ...label, sceneCount: label.scenes?.length || 0 }
   }
 
   async update(id: string, updateLabelDto: UpdateLabelDto) {
-    const label = await this.findOne(id);
-    Object.assign(label, updateLabelDto);
-    return this.labelRepository.save(label);
+    const label = await this.labelRepository.findOne({ where: { id } })
+    if (!label) {
+      throw new NotFoundException(`Label with ID ${id} not found`)
+    }
+    Object.assign(label, updateLabelDto)
+    return this.labelRepository.save(label)
   }
 
   async remove(id: string) {
-    const label = await this.findOne(id);
-    return this.labelRepository.remove(label);
+    const label = await this.labelRepository.findOne({ where: { id } })
+    if (!label) {
+      throw new NotFoundException(`Label with ID ${id} not found`)
+    }
+    return this.labelRepository.remove(label)
   }
 }
